# docs/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: HH Plus E-Commerce API
  description: HH Plus E-Commerce REST API 문서
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
...
paths:
  /users/{userId}/point:
    get:
      tags:
        - 잔액
      summary: 잔액 조회
      description: 유저의 잔액을 조회합니다.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 성공적으로 잔액을 조회했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserPoint'
              example:
                status: "SUCCESS"
                message: "성공"
                data:
                  id: 1
                  userId: 1000
                  amount: 50000
        '400':
          description: 유저가 유효하지 않습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
              example:
                status: "USER_WITHDRAWN"
                message: "탈퇴한 유저입니다."
                data: null
        '404':
          description: 유저를 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
              example:
                status: "USER_NOT_FOUND"
                message: "해당 사용자를 찾을 수 없습니다."
                data: null
  /users/{userId}/point/charge:
    post:
      tags:
        - 잔액
      summary: 잔액 충전
      description: 유저의 잔액을 충전합니다.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPointChargeRequest'
            example:
              amount: 10000
      responses:
        '200':
          description: 성공적으로 잔액을 충전했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserPoint'
              example:
                status: "SUCCESS"
                message: "성공"
                data:
                  id: 1
                  userId: 1000
                  amount: 60000
        '400':
          description: 충전 금액이 유효하지 않거나 최대 잔액을 초과합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
              examples:
                최소금액미만:
                  value:
                    status: "INVALID_CHARGE_AMOUNT"
                    message: "충전 금액은 최소 100원 이상이어야 합니다."
                    data: null
                최대금액초과:
                  value:
                    status: "INVALID_CHARGE_AMOUNT"
                    message: "충전 금액은 최대 1,000,000원 이하여야 합니다."
                    data: null
                최대잔액초과:
                  value:
                    status: "EXCEEDED_MAX_BALANCE"
                    message: "충전시 최대 잔액이 1,000,000원을 초과할 수 없습니다."
                    data: null
        '404':
          description: 유저를 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
              example:
                status: "USER_NOT_FOUND"
                message: "해당 사용자를 찾을 수 없습니다."
                data: null
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: object
          nullable: true
    ApiResponseError:
      type: object
      properties:
        status:
          type: string
          description: 에러 상태 코드
          enum:
            # 공통 에러 상태
            - SUCCESS
            - FAILURE
            - ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND

            # 유저/잔액 관련 에러 상태
            - USER_NOT_FOUND
            - USER_WITHDRAWN
            - INVALID_CHARGE_AMOUNT
            - EXCEEDED_MAX_BALANCE
            - INSUFFICIENT_BALANCE

            # 상품 관련 에러 상태
            - PRODUCT_NOT_FOUND
            - INSUFFICIENT_STOCK

            # 쿠폰 관련 에러 상태
            - COUPON_NOT_FOUND
            - COUPON_EXPIRED
            - COUPON_ALREADY_USED
            - COUPON_ISSUE_TIME_NOT_ARRIVED
            - COUPON_OUT_OF_STOCK

            # 주문 관련 에러 상태
            - ORDER_NOT_FOUND
            - INVALID_ORDER_ITEMS
        message:
          type: string
        data:
          type: object
          nullable: true

    # 잔액 관련 스키마
    UserPoint:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
    UserPointChargeRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          format: int64
          description: 충전할 금액
    ApiResponseUserPoint:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserPoint'

    # 상품 관련 스키마
...